fix(server): keep worker control service boundary clean
